onboarding_extraction_sequential/requirements.txt
```

---

### `requirements.txt`
```
camelot-py[cv]
pdfplumber
pandas
openpyxl
ghostscript